Update README.md
Global Insure, a major insurance provider, processes a large volume of claims every year. A significant portion of these claims are found to be fraudulent, resulting in financial losses and inefficient resource utilization. The current fraud detection process is primarily manual and reactive, leading to late identification of fraudulent activities. The objective is to leverage historical claim and customer data to build a predictive model that can proactively flag potentially fraudulent claims early in the process.